Fix: Enhance account management by adding role-based access control in accounts controller and introducing auth middleware

+++ server/src/endpoints/v1/accounts/accounts.controller.ts
@@ -43,13 +43,11 @@ class AccountsController {
 	async deleteAccount(request: FastifyRequest, reply: FastifyReply) {
 		const { id } = request.params as { id: string };
 
-		const account = await this.service.deleteAccount(id);
-
-		if (!account) {
-			throw new HttpException(HttpStatus.NOT_FOUND, 'Account not found');
+		if(request.fastifyUser.role === 'user' && request.fastifyUser.device_id !== id) {
+			throw new HttpException(HttpStatus.FORBIDDEN, 'You are not allowed to delete this account')
 		}
 
-		return reply.send(account);
+		return await this.service.deleteAccount(id);
 	}
 
 	async deleteDevice(request: FastifyRequest<{ Params: { deviceId: string, id: string } }>, reply: FastifyReply) {
@@ -62,18 +60,21 @@ class AccountsController {
 	async getAccountById(request: FastifyRequest<{ Params: { id: string } }>, reply: FastifyReply) {
 		const { id } = request.params;
 
-		const account = await this.service.getAccountById(id);
-
-		if (!account) {
-			throw new HttpException(HttpStatus.NOT_FOUND, 'Account not found');
+		if(request.fastifyUser.role === 'user' && request.fastifyUser.device_id !== id) {
+			throw new HttpException(HttpStatus.UNAUTHORIZED, 'You are not authorized to execute this action')
 		}
 
-		reply.send(account);
+		return await this.service.getAccountById(id);
 	}
 
 	async getAccounts(request: FastifyRequest, reply: FastifyReply) {
-		const accounts = await this.service.getAccounts();
-		reply.send(accounts);
+
+		const permitedRoles = ['admin', 'owner'];
+		if(!permitedRoles.includes(request.fastifyUser.role)) {
+			throw new HttpException(HttpStatus.FORBIDDEN, 'You are not allowed to execute this action')
+		}
+
+		return await this.service.getAccounts();
 	}
 
 	async mergeDevice(request: FastifyRequest<{ Params: { deviceId: string, id: string } }>, reply: FastifyReply) {
@@ -83,22 +84,42 @@ class AccountsController {
 		return await this.service.mergeDevice(id, deviceId);
 	}
 
-	// Todo: Needs to be protected
 	async toggleFavoriteLine(request: FastifyRequest<{ Params: { id: string, line_id: string } }>, reply: FastifyReply) {
 		const { id, line_id } = request.params;
+
+		if(request.fastifyUser?.role === 'user' && request.fastifyUser?.device_id !== id) {
+			throw new HttpException(HttpStatus.FORBIDDEN, 'You are not allowed to update this account')
+		}
 
 		return await this.service.toggleFavoriteLine(id, line_id);
 	}
 
-	// Todo: Needs to be protected
 	async toggleFavoriteStop(request: FastifyRequest<{ Params: { id: string, stop_id: string } }>, reply: FastifyReply) {
 		const { id, stop_id } = request.params;
+
+		if(request.fastifyUser?.role === 'user' && request.fastifyUser?.device_id !== id) {
+			throw new HttpException(HttpStatus.FORBIDDEN, 'You are not allowed to update this account')
+		}
 
 		return await this.service.toggleFavoriteStop(id, stop_id);
 	}
 
 	async updateAccount(request: FastifyRequest< { Params: { id: string }, Body: IAccount }>, reply: FastifyReply) {
 		const { id } = request.params;
+		
+		// If the user is not an owner, it should not be possible to change the role
+		console.log("fastifyUser", request.fastifyUser);
+		if(request.fastifyUser?.role !== 'owner') {
+			request.body.role = undefined;
+		}
+
+		// If the user is not admin/owner cant update the account
+		// If account does not belong to the user, it should not be possible to update the account
+		if(request.body.role === 'user' && request.fastifyUser.device_id !== id) {
+			throw new HttpException(HttpStatus.FORBIDDEN, 'You are not allowed to update this account')
+
+		}
+
 		return await this.service.updateAccount(id, request.body as IAccount);
 	}
 }
